Add note about error in React 18
I just wasted a lot of time setting up new project trying to figure out where the error was coming from, thought I was setting up createRoot wrong, then thought maybe it was a suspense thing, then thought it was an r3f thing, then finally figured out it was leva. Might be a good idea to add to the  readme/getting started that you will get an error in react 18, and how to fix. Great project and I appreciate it though!